Finished all security features, fixed minor bugs related to friend names in friend requests remaining after already requesting them (i.e., you can still see their name but can't request a second time until first has been accepted or rejected. Also started working on UI. Finished UI for home page.

+++ client/src/components/dashboard/friendsList/Friend.js
@@ -49,8 +49,11 @@ const Friend = ({ friendId, friendshipId }) => {
       const friendship = friends_friends.find(
         (friendship) => friendship.user_id === friendId && friendship.friend_id === user_id
       );
-      const second_friendship_id = friendship.id;
-      friendsAPI.deleteFriendship(second_friendship_id);
+      // We are first one to delete friend
+      if (friendship) {
+        const second_friendship_id = friendship.id;
+        friendsAPI.deleteFriendship(second_friendship_id);
+      }
     }
   };
